+++ src/test/java/com/renomad/minum/web/FakeRequest.java
@@ -37,4 +37,5 @@ public class FakeRequest implements IRequest {
     public Iterable<StreamingMultipartPartition> getMultipartIterable() {
         return null;
     }
+
 }